backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
mysqlclient
celery
redis
scrapy
selenium
webdriver-manager
beautifulsoup4
undetected-chromedriver
requests
django-filter
crochet
cloudscraper
playwright
requests-html
playwright-stealth>=1.0.6
lxml_html_clean
gevent
